Rename CreateIronServiceCert to CreateIronServiceCerts
