+++ ironfleet/src/CreateIronServiceCerts/CreateIronServiceCerts.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31005.135
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreateIronServiceCert", "CreateIronServiceCert.csproj", "{80249939-7D35-43CA-891A-F4C73EC6D959}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreateIronServiceCerts", "CreateIronServiceCerts.csproj", "{80249939-7D35-43CA-891A-F4C73EC6D959}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
